[#48] Restructured components folder
