+++ src/views/PostView.vue
@@ -15,8 +15,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import PostContainer from '@/components/Post/PostContainer.vue';
-import Replies from '@/components/Post/ReplyList.vue';
+import PostContainer from '@/components/post/PostContainer.vue';
+import Replies from '@/components/post/ReplyList.vue';
 import LoadingSkeleton from '@/components/LoadingSkeleton.vue';
 import { deletePost, getPost } from '@/lib/httpClient';
 import { Post } from '@/types/post';
